+++ .gitignore
@@ -1,4 +1,4 @@
 /.vs/
 /.idea
 /*.user
-/.DS_Store/
+.DS_Store/
