+++ README.md
@@ -1,17 +1,22 @@
-## Стартовый репозиторий для TheGame
+## Стартовый репозиторий
 
-Это APS.NET Core приложение с React кодом на фронтенде.
+Это ASP.NET Core приложение с React кодом на фронтенде.
 
+### Локальный запуск проекта
 
-### Для локального запуска проекта для удобной разработки фронта
+Для построения и запуска приложения
+1. перейди в директорию src
+2. выполни `dotnet build`
+3. выполни `dotnet run --environment Development`.
 
-1. Перейди в директорию src
-2. Выполнить `npm i`
-3. Выполнить `dotnet build`
-4. После успешного окончания билда, выполнить `dotnet run --environment Development`
-5. Редактировать фронт в директории `ClientApp`
+Также приложение можно запустить под отладкой в IDE обычным образом: открой sln-файл и запусти отладку.
 
-Это запустит веб-сервер в режиме разработки с горячей заменой фронтенд-кода, чтобы можно было сразу видеть результат 
-после редактирования js-кода.
+Фронт расположен в директории `ClientApp`.
 
-Можно запускать приложение из IDE, но в этом случае может не работать механизм горячей замены фронтенд-кода. 
+Веб-сервер будет запускаться в режиме разработки с горячей заменой фронтенд-кода,
+чтобы можно было сразу видеть результат  после редактирования js-кода.
+
+Для обновления бэкенда придется перестроить и запустить заново все приложение.
+
+### Troubleshooting
+Может не работать в 12 и 13 версиях node.js на MacOS и Linux. рекомендуется даунгрейд на 10 версию node.js
