--- src/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["env", "react", "stage-2"]
-}
