--- src/ClientApp/components/App/styles.css
@@ -1,9 +1,0 @@
-.root {
-    
-}
-
-.title {
-    font-size: 32px;
-    font-weight: normal;
-    text-align: center;
-}
