--- src/ClientApp/components/House/styles.css
@@ -1,9 +1,0 @@
-.root {
-    position: absolute;
-    box-sizing: border-box;
-    width: 50px;
-    height: 50px;
-    border: 1px solid #333;
-    background-color: aquamarine;
-}
-
