--- src/ClientApp/components/Person/styles.css
@@ -1,13 +1,0 @@
-.root {
-    position: absolute;
-    box-sizing: border-box;
-    width: 14px;
-    height: 14px;
-    margin-left: -7px;
-    margin-top: -7px;
-    background-color: blueviolet;
-    border: 1px solid darkmagenta;
-    border-radius: 50%;
-    transition: all 1s linear;
-}
-
