+++ src/ClientApp/public/favicon.ico
