+++ src/ClientApp/src/components/App/index.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import styles from './styles.css';
+import styles from './style.module.css';
 import Field from '../Field';
 import { DELAY, MAX_HEIGHT, MAX_WIDTH } from '../../consts/sizes';
 import { gameStateUrl, userActionUrl } from '../../consts/urls';
 import errorHandler from "../../utils/errorHandler";
 import Instruction from "../Instruction";
 
+import './base.css'
 
 export default class App extends React.Component {
     constructor() {
@@ -71,16 +72,4 @@ export default class App extends React.Component {
                 })
             })
     }
-}
-
-function getPeople(count) {
-    return Array.from({ length: count }, (_, i) => ({
-        id: i,
-        homeId: i,
-        position: {
-            x: Math.round(Math.random() * MAX_WIDTH),
-            y: Math.round(Math.random() * MAX_HEIGHT),
-        },
-        isIll: false,
-    }))
 }
