+++ src/ClientApp/src/components/App/style.module.css
@@ -1,0 +1,9 @@
+.root {
+    
+}
+
+.title {
+    font-size: 32px;
+    font-weight: normal;
+    text-align: center;
+}
