+++ src/ClientApp/src/components/House/styles.module.css
@@ -1,0 +1,9 @@
+.root {
+    position: absolute;
+    box-sizing: border-box;
+    width: 50px;
+    height: 50px;
+    border: 1px solid #333;
+    background-color: aquamarine;
+}
+
