+++ src/ClientApp/src/components/Instruction/styles.module.css
@@ -1,0 +1,32 @@
+.backdrop {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 10;
+    background-color: rgba(0, 0, 0, 0.6);
+}
+
+.modal {
+    position: absolute;
+    top: 150px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 512px;
+    padding: 30px;
+    background: black;
+    border-radius: 10px;
+    color: #f2f2f2;
+    line-height: 1.5;
+}
+
+.cross {
+    color: #f2f2f2;
+    background: none;
+    border: none;
+    font-size: 30px;
+    position: absolute;
+    right: 10px;
+    top: 10px;
+}
