+++ src/ClientApp/src/consts/sizes.js
