+++ src/Services/PersonState.cs
@@ -1,0 +1,9 @@
+namespace covidSim.Services
+{
+    internal enum PersonState
+    {
+        AtHome,
+        Walking,
+        GoingHome,
+    }
+}
