--- src/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "thegame",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "build-vendor": "webpack --config webpack.config.vendor.js",
-    "build": "webpack"
-  },
-  "devDependencies": {
-    "aspnet-webpack": "^2.0.1",
-    "aspnet-webpack-react": "^3.0.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "css-loader": "0.28.4",
-    "event-source-polyfill": "0.0.9",
-    "extract-text-webpack-plugin": "2.1.2",
-    "file-loader": "0.11.2",
-    "json-loader": "0.5.4",
-    "react-hot-loader": "3.0.0-beta.7",
-    "style-loader": "0.18.2",
-    "url-loader": "0.5.9",
-    "webpack": "2.5.1",
-    "webpack-hot-middleware": "2.18.2"
-  },
-  "dependencies": {
-    "react": "15.6.1",
-    "react-dom": "15.6.1",
-    "whatwg-fetch": "^2.0.4"
-  }
-}
