--- src/webpack.config.js
@@ -1,56 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const bundleOutputDir = './wwwroot/dist';
-
-module.exports = (env) => {
-    const isDevBuild = !(env && env.prod);
-    return [{
-        stats: { modules: false },
-        entry: { 'main': './ClientApp/boot.jsx' },
-        resolve: { extensions: ['.js', '.jsx'] },
-        output: {
-            path: path.join(__dirname, bundleOutputDir),
-            filename: '[name].js',
-            publicPath: 'dist/'
-        },
-        module: {
-            rules: [
-                { 
-                    test: /\.jsx?$/, 
-                    include: /ClientApp/, 
-                    use: 'babel-loader' 
-                }, {
-                    test: /\.css$/,
-                    loader: 'style-loader'
-                }, {
-                    test: /\.css$/,
-                    loader: 'css-loader',
-                    query: {
-                        modules: true,
-                        localIdentName: isDevBuild ? '[folder]__[local]-[hash:base64:5]' : '[hash:base64:5] ',
-                    }
-                }, { 
-                    test: /\.(png|jpg|jpeg|gif|svg)$/, 
-                    use: 'url-loader?limit=25000' 
-                }
-            ]
-        },
-        plugins: [
-            new webpack.DllReferencePlugin({
-                context: __dirname,
-                manifest: require('./wwwroot/dist/vendor-manifest.json')
-            })
-        ].concat(isDevBuild ? [
-            // Plugins that apply in development builds only
-            new webpack.SourceMapDevToolPlugin({
-                filename: '[file].map', // Remove this line if you prefer inline source maps
-                moduleFilenameTemplate: path.relative(bundleOutputDir, '[resourcePath]') // Point sourcemap entries to the original file locations on disk
-            })
-        ] : [
-            // Plugins that apply in production builds only
-            new webpack.optimize.UglifyJsPlugin(),
-            new ExtractTextPlugin('site.css')
-        ])
-    }];
-};
