--- src/webpack.config.vendor.js
@@ -1,38 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-
-module.exports = (env) => {
-    const isDevBuild = !(env && env.prod);
-    return [{
-        stats: { modules: false },
-        resolve: {
-            extensions: [ '.js' ]
-        },
-        module: {
-            rules: [
-                { test: /\.(png|woff|woff2|eot|ttf|svg)(\?|$)/, use: 'url-loader?limit=100000' },
-            ]
-        },
-        entry: {
-            vendor: ['whatwg-fetch', 'react', 'react-dom'],
-        },
-        output: {
-            path: path.join(__dirname, 'wwwroot', 'dist'),
-            publicPath: 'dist/',
-            filename: '[name].js',
-            library: '[name]_[hash]',
-        },
-        plugins: [
-            new webpack.DllPlugin({
-                path: path.join(__dirname, 'wwwroot', 'dist', '[name]-manifest.json'),
-                name: '[name]_[hash]'
-            }),
-            new webpack.DefinePlugin({
-                'process.env.NODE_ENV': isDevBuild ? '"development"' : '"production"'
-            })
-        ].concat(isDevBuild ? [] : [
-            new webpack.optimize.UglifyJsPlugin()
-        ])
-    }];
-};
