--- src/wwwroot/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>COVID-simulator</title>
-</head>
-<body>
-    <div id="react-app"></div>
-
-    <script src="/dist/vendor.js"></script>
-    <script src="/dist/main.js"></script>
-</body>
-</html>
